Fix format warnings on 64 bits OS.

+++ src/via_ums.c
@@ -187,7 +187,8 @@ viaMapMMIO(ScrnInfoPtr pScrn)
     xf86DrvMsg(pScrn->scrnIndex, X_PROBED,
                "Mapping the frame buffer at address 0x%lX with "
                "size %lu KB.\n",
-               pVia->FrameBufferBase, pVia->videoRambytes / 1024);
+               pVia->FrameBufferBase,
+              (unsigned long int) pVia->videoRambytes / 1024);
 
 #ifdef HAVE_PCIACCESS
     err = pci_device_map_range(pVia->PciInfo, pVia->FrameBufferBase,
